Move ProfilePage out of the folder and remove profile directory
